# test/swagger.yaml
openapi: 3.0.0
info:
  title: Beer Product API
  version: 1.0.0
  description: API for managing beer products.
servers:
  - url: 'https://product-service-cna-product-service.2.rahtiapp.fi/'
    description: Environment-based API URL
paths:
  /products:
    get:
      summary: Get all products
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Add a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
  /products/{id}:
    get:
      summary: Get a specific product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    put:
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated
    delete:
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Product:
      type: object
      properties:
        country:
          type: string
        category:
          type: string
        name:
          type: string
        price:
          type: number
          format: float
        description:
          type: string
        image:
          type: string
      required:
        - country
        - category
        - name
        - price
        - description
        - image
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: float